% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge.R
\name{get_edge}
\alias{get_edge}
\title{Find the links between a package's functions}
\usage{
get_edge(pkg)
}
\arguments{
\item{pkg}{a package name}
}
\value{
a character vector
\donttest{
get_function("insee")
}
}
\description{
Find the links between a package's functions
}
\details{
Find the links between a package's functions, including internal functions
}
