% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function.R
\name{get_function}
\alias{get_function}
\title{List all functions in a package}
\usage{
get_function(pkg, all_function = TRUE)
}
\arguments{
\item{pkg}{a package name}
}
\value{
a character vector
\donttest{
get_function("insee")
}
}
\description{
List all functions in a package
}
\details{
List all functions in a package, including internal functions
}
