% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizpkg.R
\name{vizpkg}
\alias{vizpkg}
\title{Visualise links between functions in a package}
\usage{
vizpkg(pkg)
}
\arguments{
\item{pkg}{a package name}
}
\value{
a visnetwork object
}
\description{
Visualise links between functions in a package
}
\details{
Visualise links between functions in a package
}
\examples{
\donttest{
vizpkg("rwebstat")
vizpkg("insee")
vizpkg("rsdmx")
}
}
